+++ back/payment/src/main/java/com/ssafy/payment/service/PaymentService.java
@@ -5,6 +5,7 @@ import com.ssafy.payment.controller.PaymentsConfirmClient;
 import com.ssafy.payment.domain.Payment;
 import com.ssafy.payment.domain.PaymentMethod;
 import com.ssafy.payment.domain.PaymentStatus;
+import com.ssafy.payment.domain.repository.PaymentFacadeRepository;
 import com.ssafy.payment.dto.request.CancelPaymentsRequest;
 import com.ssafy.payment.dto.request.ConfirmPaymentsRequest;
 import com.ssafy.payment.dto.request.RefundOrderRequest;
