+++ back/payment/src/test/java/com/ssafy/payment/PaymentServiceTest.java
@@ -9,10 +9,10 @@ import com.ssafy.payment.controller.PaymentsConfirmClient;
 import com.ssafy.payment.domain.Payment;
 import com.ssafy.payment.domain.PaymentMethod;
 import com.ssafy.payment.domain.PaymentStatus;
+import com.ssafy.payment.domain.repository.PaymentFacadeRepository;
 import com.ssafy.payment.dto.request.ConfirmPaymentsRequest;
 import com.ssafy.payment.dto.request.RefundOrderRequest;
 import com.ssafy.payment.dto.response.PaymentsResponse;
-import com.ssafy.payment.service.PaymentFacadeRepository;
 import com.ssafy.payment.service.PaymentService;
 import java.time.ZonedDateTime;
 import org.junit.jupiter.api.DisplayName;
